updating to latest changes in master

+++ spark-cassandra-connector/src/main/scala/org/apache/spark/sql/cassandra/CassandraSourceRelation.scala
@@ -1,7 +1,7 @@
 package org.apache.spark.sql.cassandra
 
 import java.net.InetAddress
-import java.util.UUID
+import java.util.{Locale, UUID}
 
 import com.datastax.spark.connector.cql.{CassandraConnector, CassandraConnectorConf, Schema}
 import com.datastax.spark.connector.rdd.partitioner.DataSizeEstimates
@@ -31,6 +31,7 @@ private[cassandra] class CassandraSourceRelation(
     tableRef: TableRef,
     userSpecifiedSchema: Option[StructType],
     filterPushdown: Boolean,
+    confirmTruncate: Boolean,
     tableSizeInBytes: Option[Long],
     connector: CassandraConnector,
     readConf: ReadConf,
@@ -53,11 +54,21 @@ private[cassandra] class CassandraSourceRelation(
 
   override def insert(data: DataFrame, overwrite: Boolean): Unit = {
     if (overwrite) {
-      connector.withSessionDo {
-        val keyspace = quote(tableRef.keyspace)
-        val table = quote(tableRef.table)
-        session => session.execute(s"TRUNCATE $keyspace.$table")
+      if (confirmTruncate) {
+        connector.withSessionDo {
+          val keyspace = quote(tableRef.keyspace)
+          val table = quote(tableRef.table)
+          session => session.execute(s"TRUNCATE $keyspace.$table")
+        }
+      } else {
+        throw new UnsupportedOperationException(
+          """You are attempting to use overwrite mode which will truncate
+          |this table prior to inserting data. If you would merely like
+          |to change data already in the table use the "Append" mode.
+          |To actually truncate please pass in true value to the option
+          |"confirm.truncate" when saving. """.stripMargin)
       }
+
     }
 
     implicit val rwf = SqlRowWriter.Factory
@@ -277,6 +288,7 @@ object CassandraSourceRelation {
       tableRef = tableRef,
       userSpecifiedSchema = schema,
       filterPushdown = options.pushdown,
+      confirmTruncate = options.confirmTruncate,
       tableSizeInBytes = tableSizeInBytes,
       connector = cassandraConnector,
       readConf = readConf,
@@ -303,15 +315,21 @@ object CassandraSourceRelation {
     val ks = tableRef.keyspace
     //Keyspace/Cluster level settings
     for (prop <- DefaultSource.confProperties) {
+      val lowerCasedProp = prop.toLowerCase(Locale.ROOT)
       val value = Seq(
+        tableConf.get(lowerCasedProp),
         tableConf.get(prop),
         sqlConf.get(s"$cluster:$ks/$prop"),
         sqlConf.get(s"$cluster/$prop"),
-        sqlConf.get(s"default/$prop")).flatten.headOption
+        sqlConf.get(s"default/$prop"),
+        sqlConf.get(prop)).flatten.headOption
       value.foreach(conf.set(prop, _))
     }
-    //Set all user properties
-    conf.setAll(tableConf -- DefaultSource.confProperties)
+    //Set all user properties not yet set
+    val SCCProps = DefaultSource
+      .confProperties
+      .flatMap(prop => Seq(prop, prop.toLowerCase(Locale.ROOT)))
+    conf.setAll(tableConf -- SCCProps)
     conf
   }
 }
